restructure the files with a spec folder and a runner + class ruby files
